compile warning fixes and remove icloud functions like setupAndQuery persistenStoreCoordinator etc

+++ Shopper/MapViewController.m
@@ -90,7 +90,7 @@
 
 -(void) actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex
 {
-    printf("Clicked button at index %d\n", buttonIndex);
+    printf("Clicked button at index %ld\n", (long)buttonIndex);
      
     switch (buttonIndex)
     {
